OP-448: Remove casperlabsnode.py file altogether from repo
